requirements.txt
# kafka and streaming
kafka-python==2.0.2

# storage / cache
redis==4.5.5

# ML and utils
numpy==1.26.0
pandas==2.2.2
scikit-learn==1.3.2
joblib==1.3.2

# log template mining
drain3==0.9.0
python-Levenshtein==0.20.9

# web frameworks
fastapi==0.99.0
uvicorn==0.22.0
flask==2.3.2
requests==2.31.0

# graph
networkx==3.1

# development (optional)
python-dotenv==1.0.0


deepdiff==6.3.0
shap==0.42.1
torch==2.0.1
networkx==3.1
scikit-learn==1.3.2
mysql-connector-python==8.1.0
